115_testing/examples/Github/wanda-master/test_requirements.txt
flake8>=3.6.0
pep8>=1.7.1
pyflakes>=2.1.0
pylint>=2.3.1
python-dotenv>=0.10.1,<1.0
pytest>=4.3.1
pytest-flake8>=1.0.4,<2.0
